+++ asdf/__init__.py
@@ -27,7 +27,17 @@ from ._version import version as __version__
 from .asdf import AsdfFile
 from .asdf import open_asdf as open  # noqa: A001
 from .config import config_context, get_config
-from .extension import AsdfExtension
 from .stream import Stream
 from .tags.core import IntegerType
 from .tags.core.external_reference import ExternalArrayReference
+
+
+def __getattr__(name):
+    if name == "AsdfExtension":
+        # defer import to only issue deprecation warning when
+        # asdf.AsdfExtension is used
+        from asdf import extension
+
+        return extension.AsdfExtension
+    msg = f"module {__name__!r} has no attribute {name!r}"
+    raise AttributeError(msg)