+++ asdf/commands/edit.py
@@ -190,8 +190,8 @@ def edit(path):
     """
     # Extract the YAML portion of the original file:
     with generic_io.get_file(path, mode="r") as fd:
-        if util.get_file_type(fd) != util.FileType.ASDF:
-            print(f"Error: '{path}' is not an ASDF file.")  # noqa: T201
+        if fd.peek(len(constants.ASDF_MAGIC)) != constants.ASDF_MAGIC:
+            print(f"Error: '{path}' is not an ASDF file.") # noqa: T201
             return 1
 
         original_content, available_bytes, contains_blocks = read_yaml(fd)