+++ CHANGES.rst
@@ -32,6 +32,7 @@ The ASDF Standard is at v1.6.0
   get_cached_asdf_extension_list in asdf.extension [#1409]
 - move asdf.types.format_tag to asdf.testing.helpers.format_tag [#1433]
 - Deprecate AsdfExtenion, AsdfExtensionList, BuiltinExtension [#1429]
+- Add AsdfDeprecationWarning to asdf_extensions entry point [#1361]
 
 2.14.3 (2022-12-15)
 -------------------