+++ CHANGES.rst
@@ -35,6 +35,7 @@ The ASDF Standard is at v1.6.0
 - Add AsdfDeprecationWarning to asdf_extensions entry point [#1361]
 - Deprecate asdf.tests.helpers [#1440]
 - respect umask when determining file permissions for written files [#1451]
+- add upper pins to dependencies based on current and expected working versions [#1461]
 
 2.14.3 (2022-12-15)
 -------------------